# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TartanHacks Backend",
    "version": "0.0.1"
  },
  "basePath": "/",
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-access-token"
      }
    }
  },
  "paths": {
...
    "/requests/team/{teamId}": {
      "get": {
        "summary": "View incoming and outgoing requests for a team",
        "tags": [
          "Team Requests Module"
        ],
        "description": "View incoming and outgoing requests for a team. Access - User",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success."
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
...
    "/schedule": {
      "post": {
        "summary": "Add new schedule item",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "tags": [
          "Schedule Module"
        ],
        "description": "Creates new schedule item. Access - Admin.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "startTime": {
                    "type": "number"
                  },
                  "endTime": {
                    "type": "number"
                  },
                  "location": {
                    "type": "string"
                  },
                  "lat": {
                    "type": "number"
                  },
                  "lng": {
                    "type": "number"
                  },
                  "platform": {
                    "type": "string"
                  },
                  "platformUrl": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success."
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
    "/schedule/{id}": {
      "patch": {
        "summary": "Edit schedule item",
        "tags": [
          "Schedule Module"
        ],
        "description": "Edit existing schedule item information. All body parameters are optional. If unspecified, the parameters are not updated. Access - Admin.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Schedule Item ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "startTime": {
                    "type": "number"
                  },
                  "endTime": {
                    "type": "number"
                  },
                  "location": {
                    "type": "string"
                  },
                  "lat": {
                    "type": "number"
                  },
                  "lng": {
                    "type": "number"
                  },
                  "platform": {
                    "type": "string"
                  },
                  "platformUrl": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success."
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      },
      "get": {
        "summary": "Get Schedule item by ID",
        "tags": [
          "Schedule Module"
        ],
        "description": "Get a single schedule item by iD. Access - Public.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Schedule Item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Success."
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      },
      "delete": {
        "summary": "Delete Schedule item by ID",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "tags": [
          "Schedule Module"
        ],
        "description": "Delete schedule item by specifying ID. Access - Admin.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Schedule Item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Success."
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
    "/schedule/": {
      "get": {
        "summary": "Get Schedule items",
        "tags": [
          "Schedule Module"
        ],
        "description": "Get all schedule items. Access - Public.",
        "responses": {
          "200": {
            "description": "Success."
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
...
    "/teams": {
      "post": {
        "summary": "List all existing teams",
        "tags": [
          "Teams Module"
        ],
        "description": "List all existing teams. Access - Admin",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success."
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
...
  },
...
}